--- src/dra_psp/10340.c
@@ -1,28 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-#include "../dra/dra.h"
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/10340", func_psp_090ECCC0);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/10340", func_psp_090ECD58);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/10340", func_psp_090ECD78);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/10340", MenuDrawAlucardCloakPreview);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/10340", func_psp_090ECF20);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/10340", func_800F6568);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/10340", func_800F6618);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/10340", func_psp_090ED058);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/10340", MenuDrawChar);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/10340", MenuDrawStr);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/10340", MenuDrawInt);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/10340", func_psp_090ED248);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/10340", func_psp_090ED2E8);
